Refactor linting function
