+++ lib/lintPatternLab.js
@@ -1,62 +1,37 @@
-'use strict';
-
-const PluginError = require('plugin-error');
-const { Transform } = require('stream');
+const { promisify } = require('util');
+const fs = require('fs');
+const glob = require('glob');
 const path = require('path');
-const { existsSync } = require('fs');
-const log = require('fancy-log');
 
-function lintPatternLab() {
-  const transform = (file, enc, cb) => {
-    const fileName = path.basename(file.path, '.twig');
-    let errorCount = 0;
-    if (fileName.charAt(0) === '_') {
-      cb(null, file);
-      return;
-    }
-    const markdownPath = path.join(path.dirname(file.path), `${fileName}.md`);
-    if (!existsSync(markdownPath)) {
-      log.error(
-        `Missing markdown file for ${path.relative(process.cwd(), file.path)}`
+const asyncGlob = promisify(glob);
+const stat = promisify(fs.stat);
+
+function lintPatternLab(options) {
+  return asyncGlob('source/_patterns/*/*/**/*.twig')
+    .then(matchedFiles => {
+      const lintedFiles = [];
+      matchedFiles.forEach(file =>
+        lintedFiles.push(lintPatternLab.lintFile(file))
       );
-      errorCount++;
-    }
-    file.patternLabErrorCount = errorCount;
-    cb(null, file);
-  };
-  return new Transform({
-    objectMode: true,
-    transform,
-  });
+      return Promise.all(lintedFiles);
+    })
+    .then(data => data.filter(item => typeof item !== 'undefined'));
 }
 
-lintPatternLab.results = function(action) {
-  let totalErrors = 0;
-  const transform = (file, enc, cb) => {
-    if (typeof file.patternLabErrorCount !== 'undefined') {
-      totalErrors += file.patternLabErrorCount;
+lintPatternLab.lintFile = function(file) {
+  return new Promise((resolve, reject) => {
+    const fileName = path.basename(file, '.twig');
+    if (fileName.charAt(0) === '_') {
+      resolve();
     }
-    cb(null, file);
-  };
-  const flush = done => {
-    action(totalErrors);
-    done();
-  };
-  return new Transform({
-    objectMode: true,
-    transform,
-    flush,
-  });
-};
-
-lintPatternLab.failAfterErrors = () => {
-  return lintPatternLab.results(totalErrorCount => {
-    if (totalErrorCount > 0) {
-      throw new PluginError('lintPatternLab', {
-        name: 'Pattern Lab Linting Error',
-        message: `Failed with ${totalErrorCount} error(s)`,
-      });
-    }
+    const markdownPath = path.join(path.dirname(file), `${fileName}.md`);
+    return stat(markdownPath)
+      .then(() => resolve())
+      .catch(() =>
+        resolve(
+          `Missing markdown file for ${path.relative(process.cwd(), file)}`
+        )
+      );
   });
 };
