+++ package.json
@@ -33,7 +33,7 @@
     "eslint-plugin-implicit-dependencies": "^1.0.4",
     "eslint-plugin-prettier": "^3.1.1",
     "fancy-log": "^1.3.3",
-    "glob": "^7.1.4",
+    "glob": "^7.1.6",
     "gulp": "^4.0.2",
     "gulp-babel": "^8.0.0",
     "gulp-postcss": "^8.0.0",
